Add files via upload
This project demonstrates the end-to-end process of building, training, and deploying a logistic regression model to predict Titanic survival outcomes based on factors like age, gender, class, and fare. The dataset is preprocessed by handling missing values and encoding categorical variables, followed by feature scaling to prepare it for model training. After training, the logistic regression model is evaluated using accuracy, precision, recall, F1-score, and ROC-AUC score. The project includes a web application built with Streamlit, allowing users to input passenger data and predict whether a passenger survived or not, as well as the survival probability. The application, along with the model and necessary code, is hosted on GitHub, providing an accessible solution for deploying machine learning models in real-world scenarios.